Trait janus_messages::codec::Encode

pub trait Encode {
    // Required method
    fn encode(&self, bytes: &mut Vec<u8, Global>);

    // Provided methods
    fn get_encoded(&self) -> Vec<u8, Global> { ... }
    fn encoded_len(&self) -> Option<usize> { ... }
}

Describes how to encode objects into a byte sequence.

Required Methods

fn encode(&self, bytes: &mut Vec<u8, Global>)

Append the encoded form of this object to the end of bytes, growing the vector as needed.

Provided Methods

fn get_encoded(&self) -> Vec<u8, Global>

Convenience method to encode a value into a new Vec<u8>.

fn encoded_len(&self) -> Option<usize>

Returns an optional hint indicating how many bytes will be required to encode this value, or None by default.