Module janus_messages::taskprov

Data structures as defined in extension draft-wang-ppm-dap-taskprov.

Structs

DpConfig

Parameters for Differential Privacy. This is mostly unspecified at the moment. See draft-irtf-cfrg-vdaf/#94 for discussion.

QueryConfig

All properties that batches for a task must have. Properties are as defined in DAP1.

TaskConfig

Defines all parameters necessary to configure an aggregator with a new task. Provided by taskprov participants in all requests incident to task execution.

VdafConfig

Describes all VDAF parameters.

Enums

DpMechanism

Query

A query type and its associated parameter(s).

VdafType