Crate janus_messages

Messages defined by the Distributed Aggregation Protocol with serialization and deserialization support.

Modules

codec

Support for encoding and decoding messages to or from the TLS wire encoding, as specified in RFC 8446, Section 3.

problem_type

query_type

Structs

AggregateShare

DAP protocol message representing a helper’s response to the leader’s request to provide an encrypted aggregate of its share of data for a given batch interval.

AggregateShareAad

DAP message representing the additional associated data for an aggregate share encryption operation.

AggregateShareReq

DAP protocol message representing a request from the leader to a helper to provide an encrypted aggregate of its share of data for a given batch interval.

AggregationJobContinueReq

DAP protocol message representing a request to continue an aggregation job.

AggregationJobId

DAP protocol message representing an identifier for an aggregation job.

AggregationJobInitializeReq

DAP protocol message representing an aggregation job initialization request from leader to helper.

AggregationJobResp

DAP protocol message representing the response to an aggregation job initialization or continuation request.

AggregationJobRound

Type representing the round of an aggregation job.

BatchId

DAP protocol message representing an ID uniquely identifying a batch, for fixed-size tasks.

BatchSelector

DAP protocol message identifying a batch of interest.

Collection

DAP protocol message representing a leader’s response to the collector’s request to provide aggregate shares for a given query.

CollectionJobId

DAP protocol message representing an identifier for a collection.

CollectionReq

DAP protocol message representing a request from the collector to the leader to provide aggregate shares for a given batch.

Duration

DAP protocol message representing a duration with a resolution of seconds.

Extension

DAP protocol message representing an arbitrary extension included in a client report.

HpkeCiphertext

DAP protocol message representing an HPKE ciphertext.

HpkeConfig

DAP protocol message representing an HPKE config.

HpkeConfigId

DAP protocol message representing an identifier for an HPKE config.

HpkeConfigList

DAP protocol message representing a list of HPKE configurations.

HpkePublicKey

DAP protocol message representing an HPKE public key.

InputShareAad

DAP message representing the additional associated data for an input share encryption operation.

Interval

DAP protocol message representing a half-open interval of time with a resolution of seconds; the start of the interval is included while the end of the interval is excluded.

PartialBatchSelector

DAP protocol message representing a partial batch selector, identifying a batch of interest in cases where some query types can infer the selector.

PlaintextInputShare

DAP protocol message representing the plaintext of an input share.

PrepareStep

DAP protocol message representing the result of a preparation step in a VDAF evaluation.

Query

Represents a query for a specific batch identifier, received from a Collector as part of the collection flow.

Report

DAP protocol message representing a client report.

ReportId

DAP protocol message representing an ID uniquely identifying a client report.

ReportIdChecksum

Checksum over DAP report IDs, defined in 4.4.4.3.

ReportMetadata

DAP protocol message representing client report metadata.

ReportShare

DAP protocol message representing one aggregator’s share of a single client report.

RoleParseError

TaskId

DAP protocol message representing an identifier for a DAP task.

Time

DAP protocol message representing an instant in time with a resolution of seconds.

Enums

Error

Errors returned by functions and methods in this module

ExtensionType

DAP protocol message representing the type of an extension included in a client report.

FixedSizeQuery

DAP protocol message representing a fixed-size query.

HpkeAeadId

DAP protocol message representing an HPKE AEAD.

HpkeKdfId

DAP protocol message representing an HPKE key derivation function.

HpkeKemId

DAP protocol message representing an HPKE key encapsulation mechanism.

PrepareStepResult

DAP protocol message representing result-type-specific data associated with a preparation step in a VDAF evaluation. Included in a PrepareStep message.

ReportShareError

DAP protocol message representing an error while preparing a report share for aggregation.

Role

DAP protocol message representing the different roles that participants can adopt.