Trait janus_messages::codec::Decode

pub trait Decode: Sized {
    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError>;

    fn get_decoded(bytes: &[u8]) -> Result<Self, CodecError> { ... }
}

Describes how to decode an object from a byte sequence.

Required Methods

fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError>

Read and decode an encoded object from bytes. On success, the decoded value is returned and bytes is advanced by the encoded size of the value. On failure, an error is returned and no further attempt to read from bytes should be made.

Provided Methods

fn get_decoded(bytes: &[u8]) -> Result<Self, CodecError>

Convenience method to get a decoded value. Returns an error if Self::decode fails, or if there are any bytes left in bytes after decoding a value.