Trait janus_messages::codec::ParameterizedDecode

pub trait ParameterizedDecode<P>: Sized {
    fn decode_with_param(
        decoding_parameter: &P,
        bytes: &mut Cursor<&[u8]>
    ) -> Result<Self, CodecError>;

    fn get_decoded_with_param(
        decoding_parameter: &P,
        bytes: &[u8]
    ) -> Result<Self, CodecError> { ... }
}

Describes how to decode an object from a byte sequence and a decoding parameter that provides additional context.

Required Methods

fn decode_with_param(
decoding_parameter: &P,
bytes: &mut Cursor<&[u8]>
) -> Result<Self, CodecError>

Read and decode an encoded object from bytes. decoding_parameter provides details of the wire encoding such as lengths of different portions of the message. On success, the decoded value is returned and bytes is advanced by the encoded size of the value. On failure, an error is returned and no further attempt to read from bytes should be made.

Provided Methods

fn get_decoded_with_param(
decoding_parameter: &P,
bytes: &[u8]
) -> Result<Self, CodecError>

Convenience method to get a decoded value. Returns an error if Self::decode_with_param fails, or if there are any bytes left in bytes after decoding a value.