Trait janus_messages::query_type::QueryType

pub trait QueryType: Clone + Debug + PartialEq + Eq + Send + Sync + 'static {
    type BatchIdentifier: Display + Debug + Clone + Hash + PartialEq + Eq + Encode + Decode + Send + Sync;
    type PartialBatchIdentifier: Debug + Clone + Hash + PartialEq + Eq + Encode + Decode + Send + Sync;
    type QueryBody: Debug + Clone + PartialEq + Eq + Encode + Decode + Send + Sync;

    const CODE: Code;

    fn partial_batch_identifier(
        batch_identifier: &Self::BatchIdentifier
    ) -> &Self::PartialBatchIdentifier;
    fn batch_identifier_for_collection(
        query: &Query<Self>,
        collect_resp: &Collection<Self>
    ) -> Self::BatchIdentifier;
}

QueryType represents a DAP query type. This is a task-level configuration setting which determines how individual client reports are grouped together into batches for collection.

Required Associated Types

type BatchIdentifier: Display + Debug + Clone + Hash + PartialEq + Eq + Encode + Decode + Send + Sync

The type of a batch identifier.

type PartialBatchIdentifier: Debug + Clone + Hash + PartialEq + Eq + Encode + Decode + Send + Sync

The type of a batch identifier as it appears in a PartialBatchSelector. Will be either the same type as BatchIdentifier, or ().

type QueryBody: Debug + Clone + PartialEq + Eq + Encode + Decode + Send + Sync

The type of the body of a Query for this query type.

Required Associated Constants

const CODE: Code

The Code associated with this query type.

Required Methods

fn partial_batch_identifier(
batch_identifier: &Self::BatchIdentifier
) -> &Self::PartialBatchIdentifier

Computes the PartialBatchIdentifier corresponding to the given BatchIdentifier.

fn batch_identifier_for_collection(
query: &Query<Self>,
collect_resp: &Collection<Self>
) -> Self::BatchIdentifier

Retrieves the batch identifier associated with an ongoing collection.